+++ src/components/[guild]/EditButton/hooks/useEdit.ts
@@ -1,7 +1,7 @@
 import useShowErrorToast from "components/create-guild/hooks/useShowErrorToast"
 import { useGuild } from "components/[guild]/Context"
 import usePersonalSign from "hooks/usePersonalSign"
-import useSubmitMachine from "hooks/useSubmitMachine"
+import useSubmit from "hooks/useSubmit"
 import useToast from "hooks/useToast"
 import { useSWRConfig } from "swr"
 
@@ -24,7 +24,7 @@ const useEdit = (onClose: () => void) => {
       body: JSON.stringify({ addressSignedMessage, ...data }),
     })
 
-  return useSubmitMachine<Data, any>(submit, {
+  return useSubmit<Data, any>(submit, {
     onSuccess: () => {
       toast({
         title: `Guild successfully updated!`,
