+++ src/hooks/useSubmit/index.ts
@@ -1,0 +1,3 @@
+import useSubmit from "./useSubmit"
+
+export default useSubmit
