--- src/hooks/useSubmitMachine/index.ts
@@ -1,3 +1,0 @@
-import useSubmitMachine from "./useSubmitMachine"
-
-export default useSubmitMachine
